+++ lib/file/file-process/index.js
@@ -1,2 +1,3 @@
 export * from './get-content';
 export * from './check-file';
+export * from './utils';
